+++ src/tests/test_integration.py
@@ -660,7 +660,7 @@ def test_google_convention(env):
     assert 'D409' not in out
     assert 'D410' not in out
     assert 'D412' in out
-    assert 'D413' in out
+    assert 'D413' not in out
     assert 'D414' in out
     assert 'D417' in out